third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/forms/constraints/form-validation-reportValidity-expected.txt

ValidationMessageClient: main-message=Please shorten this text to 4 characters or less (you are currently using 5 characters). sub-message=
ValidationMessageClient: main-message=Please match the requested format. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
ValidationMessageClient: main-message=Please shorten this text to 4 characters or less (you are currently using 5 characters). sub-message=
ValidationMessageClient: main-message=Please match the requested format. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
ValidationMessageClient: main-message=Please shorten this text to 4 characters or less (you are currently using 5 characters). sub-message=
ValidationMessageClient: main-message=Please match the requested format. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
ValidationMessageClient: main-message=Please shorten this text to 4 characters or less (you are currently using 5 characters). sub-message=
ValidationMessageClient: main-message=Please match the requested format. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
ValidationMessageClient: main-message=Please shorten this text to 20 characters or less (you are currently using 21 characters). sub-message=
ValidationMessageClient: main-message=Please match the requested format. sub-message=
ValidationMessageClient: main-message=Please enter a URL. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
CONSOLE WARNING: The specified value "a" is not a valid email address.
CONSOLE WARNING: The specified value "a" is not a valid email address.
ValidationMessageClient: main-message=Please shorten this text to 10 characters or less (you are currently using 15 characters). sub-message=
ValidationMessageClient: main-message=Please match the requested format. sub-message=
CONSOLE WARNING: The specified value "abc" is not a valid email address.
ValidationMessageClient: main-message=Please include an '@' in the email address. 'abc' is missing an '@'. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
ValidationMessageClient: main-message=Value must be 01/01/2000 or earlier. sub-message=
ValidationMessageClient: main-message=Value must be 01/01/2001 or later. sub-message=
ValidationMessageClient: main-message=Please enter a valid value. The nearest valid value is 01/01/1970. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
ValidationMessageClient: main-message=Value must be January 2000 or earlier. sub-message=
ValidationMessageClient: main-message=Value must be January 2001 or later. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
ValidationMessageClient: main-message=Value must be Week 01, 2000 or earlier. sub-message=
ValidationMessageClient: main-message=Value must be Week 01, 2001 or later. sub-message=
ValidationMessageClient: main-message=Please enter a valid value. The nearest valid value is Week 01, 1970. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
ValidationMessageClient: main-message=Value must be 12:00 PM or earlier. sub-message=
ValidationMessageClient: main-message=Value must be 12:00 PM or later. sub-message=
ValidationMessageClient: main-message=Please enter a valid value. The nearest valid value is 12:00 AM. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
ValidationMessageClient: main-message=Value must be less than or equal to 5. sub-message=
ValidationMessageClient: main-message=Value must be greater than or equal to 5. sub-message=
ValidationMessageClient: main-message=Please enter a valid value. The two nearest valid values are 2 and 4. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
ValidationMessageClient: main-message=Please check this box if you want to proceed. sub-message=
ValidationMessageClient: main-message=Please select one of these options. sub-message=
ValidationMessageClient: main-message=Please select a file. sub-message=
ValidationMessageClient: main-message=Please select an item in the list. sub-message=
ValidationMessageClient: main-message=Please fill out this field. sub-message=
This is a testharness.js-based test.
PASS [INPUT in TEXT status] no constraint 
PASS [INPUT in TEXT status] no constraint (in a form) 
FAIL [INPUT in TEXT status] suffering from being too long assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in TEXT status] suffering from being too long (in a form) 
FAIL [INPUT in TEXT status] suffering from a pattern mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in TEXT status] suffering from a pattern mismatch (in a form) 
FAIL [INPUT in TEXT status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in TEXT status] suffering from being missing (in a form) 
PASS [INPUT in SEARCH status] no constraint 
PASS [INPUT in SEARCH status] no constraint (in a form) 
FAIL [INPUT in SEARCH status] suffering from being too long assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in SEARCH status] suffering from being too long (in a form) 
FAIL [INPUT in SEARCH status] suffering from a pattern mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in SEARCH status] suffering from a pattern mismatch (in a form) 
FAIL [INPUT in SEARCH status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in SEARCH status] suffering from being missing (in a form) 
PASS [INPUT in TEL status] no constraint 
PASS [INPUT in TEL status] no constraint (in a form) 
FAIL [INPUT in TEL status] suffering from being too long assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in TEL status] suffering from being too long (in a form) 
FAIL [INPUT in TEL status] suffering from a pattern mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in TEL status] suffering from a pattern mismatch (in a form) 
FAIL [INPUT in TEL status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in TEL status] suffering from being missing (in a form) 
PASS [INPUT in PASSWORD status] no constraint 
PASS [INPUT in PASSWORD status] no constraint (in a form) 
FAIL [INPUT in PASSWORD status] suffering from being too long assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in PASSWORD status] suffering from being too long (in a form) 
FAIL [INPUT in PASSWORD status] suffering from a pattern mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in PASSWORD status] suffering from a pattern mismatch (in a form) 
FAIL [INPUT in PASSWORD status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in PASSWORD status] suffering from being missing (in a form) 
PASS [INPUT in URL status] no constraint 
PASS [INPUT in URL status] no constraint (in a form) 
FAIL [INPUT in URL status] suffering from being too long assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in URL status] suffering from being too long (in a form) 
FAIL [INPUT in URL status] suffering from a pattern mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in URL status] suffering from a pattern mismatch (in a form) 
FAIL [INPUT in URL status] suffering from a type mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in URL status] suffering from a type mismatch (in a form) 
FAIL [INPUT in URL status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in URL status] suffering from being missing (in a form) 
PASS [INPUT in EMAIL status] no constraint 
PASS [INPUT in EMAIL status] no constraint (in a form) 
FAIL [INPUT in EMAIL status] suffering from being too long assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in EMAIL status] suffering from being too long (in a form) 
FAIL [INPUT in EMAIL status] suffering from a pattern mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in EMAIL status] suffering from a pattern mismatch (in a form) 
FAIL [INPUT in EMAIL status] suffering from a type mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in EMAIL status] suffering from a type mismatch (in a form) 
FAIL [INPUT in EMAIL status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in EMAIL status] suffering from being missing (in a form) 
FAIL [INPUT in DATETIME status] The datetime type must be supported. assert_equals: The datetime type should be supported. expected "datetime" but got "text"
PASS [INPUT in DATE status] no constraint 
PASS [INPUT in DATE status] no constraint (in a form) 
FAIL [INPUT in DATE status] suffering from an overflow assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in DATE status] suffering from an overflow (in a form) 
FAIL [INPUT in DATE status] suffering from an underflow assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in DATE status] suffering from an underflow (in a form) 
FAIL [INPUT in DATE status] suffering from a step mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in DATE status] suffering from a step mismatch (in a form) 
FAIL [INPUT in DATE status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in DATE status] suffering from being missing (in a form) 
PASS [INPUT in MONTH status] no constraint 
PASS [INPUT in MONTH status] no constraint (in a form) 
FAIL [INPUT in MONTH status] suffering from an overflow assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in MONTH status] suffering from an overflow (in a form) 
FAIL [INPUT in MONTH status] suffering from an underflow assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in MONTH status] suffering from an underflow (in a form) 
FAIL [INPUT in MONTH status] suffering from a step mismatch assert_true: The invalid event should be fired. expected true got false
FAIL [INPUT in MONTH status] suffering from a step mismatch (in a form) assert_false: The reportValidity method of the element's form owner should return false. expected false got true
FAIL [INPUT in MONTH status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in MONTH status] suffering from being missing (in a form) 
PASS [INPUT in WEEK status] no constraint 
PASS [INPUT in WEEK status] no constraint (in a form) 
FAIL [INPUT in WEEK status] suffering from an overflow assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in WEEK status] suffering from an overflow (in a form) 
FAIL [INPUT in WEEK status] suffering from an underflow assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in WEEK status] suffering from an underflow (in a form) 
FAIL [INPUT in WEEK status] suffering from a step mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in WEEK status] suffering from a step mismatch (in a form) 
FAIL [INPUT in WEEK status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in WEEK status] suffering from being missing (in a form) 
PASS [INPUT in TIME status] no constraint 
PASS [INPUT in TIME status] no constraint (in a form) 
FAIL [INPUT in TIME status] suffering from an overflow assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in TIME status] suffering from an overflow (in a form) 
FAIL [INPUT in TIME status] suffering from an underflow assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in TIME status] suffering from an underflow (in a form) 
FAIL [INPUT in TIME status] suffering from a step mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in TIME status] suffering from a step mismatch (in a form) 
FAIL [INPUT in TIME status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in TIME status] suffering from being missing (in a form) 
FAIL [INPUT in NUMBER status] suffering from an overflow assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in NUMBER status] suffering from an overflow (in a form) 
FAIL [INPUT in NUMBER status] suffering from an underflow assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in NUMBER status] suffering from an underflow (in a form) 
FAIL [INPUT in NUMBER status] suffering from a step mismatch assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in NUMBER status] suffering from a step mismatch (in a form) 
FAIL [INPUT in NUMBER status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in NUMBER status] suffering from being missing (in a form) 
PASS [INPUT in CHECKBOX status] no constraint 
PASS [INPUT in CHECKBOX status] no constraint (in a form) 
FAIL [INPUT in CHECKBOX status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in CHECKBOX status] suffering from being missing (in a form) 
PASS [INPUT in RADIO status] no constraint 
PASS [INPUT in RADIO status] no constraint (in a form) 
FAIL [INPUT in RADIO status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in RADIO status] suffering from being missing (in a form) 
PASS [INPUT in FILE status] no constraint 
PASS [INPUT in FILE status] no constraint (in a form) 
FAIL [INPUT in FILE status] suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [INPUT in FILE status] suffering from being missing (in a form) 
PASS [select]  no constraint 
PASS [select]  no constraint (in a form) 
FAIL [select]  suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [select]  suffering from being missing (in a form) 
PASS [textarea]  no constraint 
PASS [textarea]  no constraint (in a form) 
FAIL [textarea]  suffering from being missing assert_true: The invalid event should be fired. expected true got false
PASS [textarea]  suffering from being missing (in a form) 
Harness: the test ran to completion.
